[Cursor] Update audio processing system to include transformers library and enhance WorkflowState model. Added the transformers dependency to the project and updated the WorkflowState model to include new fields for audio content and critique handling. Improved error handling and retry mechanisms in the Coordinator and MCPCodeExecutor classes for better robustness. These changes enhance the audio processing workflow and ensure compatibility with the AudioLDM model for audio generation.

+++ audio_editor/audio_tools/ai.py
@@ -151,7 +151,7 @@ def AUDIO_GENERATE(text: str, filename: str, audio_length_in_s: float = 5.0, gui
         try:
             _audio_gen_pipe = diffusers.AudioLDMPipeline.from_pretrained(
                 "cvssp/audioldm-s-full-v2", 
-                torch_dtype=torch.float16
+                torch_dtype=torch.float32  # Use float32 for consistent dtype
             )
             _audio_gen_pipe = _audio_gen_pipe.to("cuda" if torch.cuda.is_available() else "cpu")
         except Exception as e:
